+++ lib/browser/browser_core.ts
@@ -1,14 +1,16 @@
 import path from 'path';
 import querystring from 'querystring';
 
-import { stringifyLogText } from './puppeteer_wrapper/puppeteer_utils';
+import { stringifyLogText } from '../puppeteer_wrapper/puppeteer_utils';
 import WendigoConfig from '../../config';
 import DomElement from '../models/dom_element';
 import { FatalError, InjectScriptError } from '../errors';
 import { FinalBrowserSettings, OpenSettings } from '../types';
-import PuppeteerPage from './puppeteer_wrapper/puppeteer_page';
-import { ViewportOptions, ConsoleMessage, Page, Response, Frame } from './puppeteer_wrapper/puppeteer_types';
+import PuppeteerPage from '../puppeteer_wrapper/puppeteer_page';
+import { ViewportOptions, ConsoleMessage, Page, Response, Frame, BrowserContext, Target } from '../puppeteer_wrapper/puppeteer_types';
 import FailIfNotLoaded from '../decorators/fail_if_not_loaded';
+import PuppeteerContext from '../puppeteer_wrapper/puppeteer_context';
+import OverrideError from '../decorators/override_error';
 
 const injectionScriptsPath = WendigoConfig.injectionScripts.path;
 const injectionScripts = WendigoConfig.injectionScripts.files;
@@ -36,25 +38,42 @@ export default abstract class BrowserCore {
     public initialResponse: Response | null;
 
     protected _page: PuppeteerPage;
+    protected _context: PuppeteerContext;
     protected originalHtml?: string;
     protected settings: FinalBrowserSettings;
 
     private _loaded: boolean;
-    private disabled: boolean;
-    private components: Array<string>;
-    private cache: boolean;
+    private _disabled: boolean;
+    private _components: Array<string>;
+    private _cache: boolean;
+    private _openSettings: OpenSettings = defaultOpenOptions;
 
-    constructor(page: PuppeteerPage, settings: FinalBrowserSettings, components: Array<string> = []) {
+    constructor(context: PuppeteerContext, page: PuppeteerPage, settings: FinalBrowserSettings, components: Array<string> = []) {
         this._page = page;
+        this._context = context;
         this.settings = settings;
         this._loaded = false;
         this.initialResponse = null;
-        this.disabled = false;
-        this.cache = settings.cache !== undefined ? settings.cache : true;
-        this.components = components;
+        this._disabled = false;
+        this._cache = settings.cache !== undefined ? settings.cache : true;
+        this._components = components;
         if (this.settings.log) {
             this._page.on("console", pageLog);
         }
+
+        this._context.on('targetcreated', async (target: Target): Promise<void> => {
+            const createdPage = await target.page();
+            if (createdPage) {
+                const puppeteerPage = new PuppeteerPage(createdPage);
+                try {
+                    await puppeteerPage.setBypassCSP(true);
+                    if (this.settings.userAgent)
+                        await puppeteerPage.setUserAgent(this.settings.userAgent);
+                } catch (err) {
+                    // Will fail if browser is closed before finishing
+                }
+            }
+        });
 
         this._page.on('load', async (): Promise<void> => {
             if (this._loaded) {
@@ -70,8 +89,13 @@ export default abstract class BrowserCore {
     public get page(): Page {
         return this._page.page;
     }
+
+    public get context(): BrowserContext {
+        return this._context.context;
+    }
+
     public get loaded(): boolean {
-        return this._loaded && !this.disabled;
+        return this._loaded && !this._disabled;
     }
 
     public get incognito(): boolean {
@@ -79,20 +103,21 @@ export default abstract class BrowserCore {
     }
 
     public get cacheEnabled(): boolean {
-        return this.cache;
+        return this._cache;
     }
 
+    @OverrideError()
     public async open(url: string, options?: OpenSettings): Promise<void> {
         this._loaded = false;
-        options = Object.assign({}, defaultOpenOptions, options);
+        this._openSettings = Object.assign({}, defaultOpenOptions, options);
         url = this._processUrl(url);
-        await this.setCache(this.cache);
-        if (options.queryString) {
-            const qs = this._generateQueryString(options.queryString);
+        await this.setCache(this._cache);
+        if (this._openSettings.queryString) {
+            const qs = this._generateQueryString(this._openSettings.queryString);
             url = `${url}${qs}`;
         }
         try {
-            await this._beforeOpen(options);
+            await this._beforeOpen(this._openSettings);
             const response = await this._page.goto(url);
             this.initialResponse = response;
             return this._afterPageLoad();
@@ -102,6 +127,7 @@ export default abstract class BrowserCore {
         }
     }
 
+    @OverrideError()
     public async openFile(filepath: string, options: OpenSettings): Promise<void> {
         const absolutePath = path.resolve(filepath);
         try {
@@ -112,9 +138,9 @@ export default abstract class BrowserCore {
     }
 
     public async close(): Promise<void> {
-        if (this.disabled) return Promise.resolve();
-        const p = this._beforeClose();
-        this.disabled = true;
+        if (this._disabled) return Promise.resolve();
+        const p = this._beforeClose(); // Minor race condition with this._loaded if moved
+        this._disabled = true;
         this._loaded = false;
         this.initialResponse = null;
         this.originalHtml = undefined;
@@ -134,6 +160,32 @@ export default abstract class BrowserCore {
         if (resultAsElement) {
             return new DomElement(resultAsElement);
         } else return rawResult.jsonValue();
+    }
+
+    public async pages(): Promise<Array<Page>> {
+        return this._context.pages();
+    }
+
+    @OverrideError()
+    public async selectPage(index: number): Promise<void> {
+        const page = await this._context.getPage(index);
+        if (!page) throw new FatalError("selectPage", `Invalid page index "${index}".`);
+        this._page = page;
+        // TODO: Avoid reload
+        await this.page.reload(); // Required to enable bypassCSP
+        await this._beforeOpen(this._openSettings);
+        await this._afterPageLoad();
+    }
+
+    public async closePage(index: number): Promise<void> {
+        const page = await this._context.getPage(index);
+        if (!page) throw new FatalError("closePage", `Invalid page index "${index}".`);
+        await page.close();
+        try {
+            await this.selectPage(0);
+        } catch (err) {
+            this.close();
+        }
     }
 
     public setViewport(config: ViewportOptions = {}): Promise<void> {
@@ -165,13 +217,13 @@ export default abstract class BrowserCore {
                 path: scriptPath
             });
         } catch (err) {
-            return Promise.reject(new InjectScriptError("open", err));
+            return Promise.reject(new InjectScriptError("addScript", err));
         }
     }
 
     public async setCache(value: boolean): Promise<void> {
         await this._page.setCache(value);
-        this.cache = value;
+        this._cache = value;
     }
 
     protected async _beforeClose(): Promise<void> {
@@ -184,7 +236,6 @@ export default abstract class BrowserCore {
         if (this.settings.userAgent) {
             await this._page.setUserAgent(this.settings.userAgent);
         }
-
         if (this.settings.bypassCSP) {
             await this._page.setBypassCSP(true);
         }
@@ -198,7 +249,10 @@ export default abstract class BrowserCore {
             this.originalHtml = content;
             await this._addJsScripts();
         } catch (err) {
-            if (err.message === "Evaluation failed: Event") throw new InjectScriptError("open", err.message); // CSP error
+            if (err.message === "Evaluation failed: Event") {
+                const cspWarning = "This may be caused by the page Content Security Policy. Make sure the option bypassCSP is set to true in Wendigo.";
+                throw new InjectScriptError("_afterPageLoad", `Error injecting scripts. ${cspWarning}`); // CSP error
+            }
         }
         this._loaded = true;
         await this._callComponentsMethod("_afterOpen");
@@ -221,7 +275,7 @@ export default abstract class BrowserCore {
     }
 
     private async _callComponentsMethod(method: string, options?: any): Promise<void> {
-        await Promise.all(this.components.map((c) => {
+        await Promise.all(this._components.map((c) => {
             const anyThis = this as any;
             if (typeof anyThis[c][method] === 'function')
                 return anyThis[c][method](options);