+++ lib/modules/cookies/browser_cookies.ts
@@ -1,4 +1,4 @@
-import { Cookie as CookieData, SetCookie, DeleteCookie } from '../../browser/puppeteer_wrapper/puppeteer_types';
+import { Cookie as CookieData, SetCookie, DeleteCookie } from '../../puppeteer_wrapper/puppeteer_types';
 import WendigoModule from '../wendigo_module';
 import { WendigoError } from '../../errors';
 import { arrayfy } from '../../utils/utils';
@@ -6,7 +6,7 @@ import { arrayfy } from '../../utils/utils';
 export default class BrowserCookies extends WendigoModule {
     public async all(): Promise<{ [s: string]: string }> {
         const cookies = await this._page.cookies();
-        return cookies.reduce((acc, cookie): { [s: string]: string } => {
+        return cookies.reduce((acc, cookie: CookieData): { [s: string]: string } => {
             acc[cookie.name] = cookie.value;
             return acc;
         }, {} as { [s: string]: string });
@@ -38,7 +38,7 @@ export default class BrowserCookies extends WendigoModule {
     public delete(name: string | Array<string> | DeleteCookie): Promise<void> {
         if (name === undefined || name === null) throw new WendigoError("cookies.delete", "Delete cookie name missing");
 
-        if (this.isDeleteCookieInterface(name)) {
+        if (this._isDeleteCookieInterface(name)) {
             return this._page.deleteCookie(name);
         }
 
@@ -56,7 +56,7 @@ export default class BrowserCookies extends WendigoModule {
         return this.delete(cookiesList);
     }
 
-    private isDeleteCookieInterface(data: any): data is DeleteCookie {
+    private _isDeleteCookieInterface(data: any): data is DeleteCookie {
         if (data.name) return true;
         else return false;
     }