+++ lib/modules/dialog/browser_dialog.ts
@@ -14,53 +14,53 @@ const defaultOptions: DialogOptions = {
 };
 
 export default class BrowserDialog extends WendigoModule {
-    private dialogs: Array<Dialog>;
-    private options: DialogOptions;
-    private onDialogCB?: (d: Dialog) => void;
-    private lastDialog?: Dialog;
+    private _dialogs: Array<Dialog>;
+    private _options: DialogOptions;
+    private _onDialogCB?: (d: Dialog) => void;
+    private _lastDialog?: Dialog;
 
     constructor(browser: Browser) {
         super(browser);
-        this.dialogs = [];
-        this.options = Object.assign({}, defaultOptions);
+        this._dialogs = [];
+        this._options = Object.assign({}, defaultOptions);
 
         browser.page.on("dialog", (rawDialog) => {
             const newDialog = new Dialog(rawDialog);
-            this.dialogs.push(newDialog);
-            if (this.onDialogCB) {
-                this.onDialogCB(newDialog);
-                this.onDialogCB = undefined;
-                this.lastDialog = undefined;
+            this._dialogs.push(newDialog);
+            if (this._onDialogCB) {
+                this._onDialogCB(newDialog);
+                this._onDialogCB = undefined;
+                this._lastDialog = undefined;
             } else {
-                this.lastDialog = newDialog;
+                this._lastDialog = newDialog;
             }
-            if (this.options.dismissAllDialogs) newDialog.dismiss();
+            if (this._options.dismissAllDialogs) newDialog.dismiss();
         });
     }
 
     public all(): Array<Dialog> {
-        return this.dialogs;
+        return this._dialogs;
     }
 
     public clear(): void {
-        this.dialogs = [];
-        this.lastDialog = undefined;
+        this._dialogs = [];
+        this._lastDialog = undefined;
     }
 
     public waitForDialog(timeout = 500): Promise<Dialog> {
-        if (this.lastDialog) {
-            const result = Promise.resolve(this.lastDialog);
-            this.lastDialog = undefined;
+        if (this._lastDialog) {
+            const result = Promise.resolve(this._lastDialog);
+            this._lastDialog = undefined;
             return result;
         } else {
             return new Promise((resolve, reject) => {
                 const tid = setTimeout(() => {
-                    if (this.onDialogCB) {
-                        this.onDialogCB = undefined;
+                    if (this._onDialogCB) {
+                        this._onDialogCB = undefined;
                         reject(new TimeoutError("dialog.waitForDialog", "", timeout));
                     }
                 }, timeout);
-                this.onDialogCB = (dialog) => {
+                this._onDialogCB = (dialog) => {
                     clearTimeout(tid);
                     resolve(dialog);
                 };
@@ -69,14 +69,14 @@ export default class BrowserDialog extends WendigoModule {
     }
 
     public dismissLast(): Promise<void> {
-        if (this.dialogs.length === 0) return Promise.resolve();
-        return this.dialogs[this.dialogs.length - 1].dismiss();
+        if (this._dialogs.length === 0) return Promise.resolve();
+        return this._dialogs[this._dialogs.length - 1].dismiss();
     }
 
     protected async _beforeOpen(options: OpenSettings): Promise<void> {
         await super._beforeOpen(options);
         this.clear();
-        this.options = Object.assign({}, defaultOptions);
-        if ((options as any).dismissAllDialogs) this.options.dismissAllDialogs = true;
+        this._options = Object.assign({}, defaultOptions);
+        if ((options as any).dismissAllDialogs) this._options.dismissAllDialogs = true;
     }
 }